Improve CLusterInfo
