+++ extra/man/stackablectl.1
@@ -2,10 +2,11 @@
 .el .ds Aq '
 .TH stackablectl 1  "stackablectl 0.0.0-dev" 
 .SH NAME
-stackablectl
+stackablectl \- Command line tool to interact with a Stackable Data Platform
 .SH SYNOPSIS
 \fBstackablectl\fR [\fB\-l\fR|\fB\-\-log\-level\fR] [\fB\-\-no\-cache\fR] [\fB\-\-offline\fR] [\fB\-d\fR|\fB\-\-demo\-file\fR] [\fB\-s\fR|\fB\-\-stack\-file\fR] [\fB\-r\fR|\fB\-\-release\-file\fR] [\fB\-\-helm\-repo\-stable\fR] [\fB\-\-helm\-repo\-test\fR] [\fB\-\-helm\-repo\-dev\fR] [\fB\-h\fR|\fB\-\-help\fR] [\fB\-V\fR|\fB\-\-version\fR] <\fIsubcommands\fR>
 .SH DESCRIPTION
+Command line tool to interact with a Stackable Data Platform
 .SH OPTIONS
 .TP
 \fB\-l\fR, \fB\-\-log\-level\fR=\fILOG_LEVEL\fR
@@ -95,3 +96,5 @@ stackablectl\-help(1)
 Print this message or the help of the given subcommand(s)
 .SH VERSION
 v0.0.0\-dev
+.SH AUTHORS
+Stackable GmbH <info@stackable.tech>
