+++ rust/stackable-cockpit/src/engine/docker/mod.rs
