+++ rust/stackable-cockpit/src/engine/kind/config.rs
@@ -1,6 +1,6 @@
 use serde::Serialize;
 
-use crate::cluster::NodeRole;
+use crate::engine::NodeRole;
 
 #[derive(Debug, Serialize)]
 #[serde(rename_all = "camelCase")]
