+++ rust/stackable-cockpit/src/lib.rs
@@ -1,6 +1,6 @@
-pub mod cluster;
 pub mod common;
 pub mod constants;
+pub mod engine;
 pub mod helm;
 pub mod platform;
 pub mod utils;
