+++ rust/stackablectl/src/args/cluster.rs
@@ -2,8 +2,8 @@ use clap::{Args, ValueEnum};
 use snafu::{ensure, ResultExt, Snafu};
 
 use stackable_cockpit::{
-    cluster::{KindCluster, KindClusterError, MinikubeCluster, MinikubeClusterError},
     constants::DEFAULT_LOCAL_CLUSTER_NAME,
+    engine::{KindCluster, KindClusterError, MinikubeCluster, MinikubeClusterError},
 };
 
 #[derive(Debug, Snafu)]
